requirements.txt
requests>=2.28.0
selenium>=4.15.0
Pillow>=9.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
urllib3>=1.26.0
